+++ packages/terra-worklist-data-grid/src/proptypes/cellShape.js
@@ -13,6 +13,10 @@ const cellShape = PropTypes.shape({
    * Boolean value indicating whether or not the column header is selectable.
    */
   isSelectable: PropTypes.bool,
+  /**
+   * Boolean value indicating whether or not the cell should render as selected.
+   */
+  isSelected: PropTypes.bool,
 });
 
 export default cellShape;