+++ packages/terra-worklist-data-grid/tests/jest/Row.test.jsx
@@ -1,6 +1,4 @@
 import React from 'react';
-/* eslint-disable-next-line import/no-extraneous-dependencies */
-import { shallowWithIntl } from 'terra-enzyme-intl';
 import Row from '../../src/subcomponents/Row';
 import Cell from '../../src/subcomponents/Cell';
 
@@ -43,7 +41,7 @@ describe('Row', () => {
   it('creates a row with the correct number of cells', () => {
     const rowData = dataFile.rows[0];
 
-    const wrapper = shallowWithIntl(
+    const wrapper = shallow(
       <Row
         rowIndex={99}
         key={12}
@@ -79,7 +77,6 @@ describe('Row', () => {
       expect(cell.props.rowIndex).toEqual(rowIndex);
       expect(cell.props.columnIndex).toEqual(cellIndex);
       expect(cell.props.isRowHeader).toEqual(isRowHeader);
-      expect(cell.props.isSelected).toEqual(false);
       expect(cell.props.isMasked).toEqual(expectedCell.isMasked);
       expect(cell.props.onCellSelect).toBeDefined();
       expect(cell.key).toEqual(`${rowId}_${columnId}`);
@@ -87,7 +84,7 @@ describe('Row', () => {
 
     const rowData = dataFile.rows[rowIndex];
 
-    const wrapper = shallowWithIntl(
+    const wrapper = shallow(
       <Row
         rowIndex={rowIndex}
         key={rowIndex}