+++ packages/terra-menu/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Unreleased
 
+## 6.76.0 - (August 31, 2023)
+
+* Changed
+  * Minor dependency version bump
+
 ## 6.75.0 - (August 10, 2023)
 
 * Added